Moved data controllers/models to a more appropriate package(DATA).
