+++ src/me/corriekay/pokegoutil/BlossomsPoGoManager.java
@@ -7,7 +7,7 @@ import javafx.scene.Scene;
 import javafx.scene.image.Image;
 import javafx.stage.Stage;
 import javafx.stage.StageStyle;
-import me.corriekay.pokegoutil.controllers.AccountController;
+import me.corriekay.pokegoutil.DATA.controllers.AccountController;
 import me.corriekay.pokegoutil.utils.Config;
 import me.corriekay.pokegoutil.utils.ui.Console;
 import me.corriekay.pokegoutil.utils.helpers.UIHelper;
