+++ src/me/corriekay/pokegoutil/DATA/controllers/InventoryController.java
@@ -1,4 +1,4 @@
-package me.corriekay.pokegoutil.controllers;
+package me.corriekay.pokegoutil.DATA.controllers;
 
 import com.pokegoapi.api.PokemonGo;
