+++ src/me/corriekay/pokegoutil/DATA/models/Inventory.java
@@ -1,4 +1,4 @@
-package me.corriekay.pokegoutil.models;
+package me.corriekay.pokegoutil.DATA.models;
 
 public class Inventory {
     // list of items and such
