+++ src/me/corriekay/pokegoutil/DATA/models/PlayerAccount.java
@@ -1,4 +1,4 @@
-package me.corriekay.pokegoutil.models;
+package me.corriekay.pokegoutil.DATA.models;
 
 public class PlayerAccount {
     //A player account contains information about the player
