+++ src/me/corriekay/pokegoutil/DATA/models/PokemonBag.java
@@ -1,4 +1,4 @@
-package me.corriekay.pokegoutil.models;
+package me.corriekay.pokegoutil.DATA.models;
 
 public class PokemonBag {
     // list of pokemons with different informations that we use to display
